fix: cross-platform compilation - 42 remaining errors in Registry and GameState

+++ GeneralsMD/Code/GameEngine/Source/Common/INI/INI.cpp
@@ -659,7 +659,7 @@ void INI::parseBool( INI* ini, void * /*instance*/, void *store, const void* /*u
 void INI::parseBitInInt32( INI *ini, void *instance, void *store, const void* userData )
 {
 	UnsignedInt* s = (UnsignedInt*)store;
-	UnsignedInt mask = (UnsignedInt)userData;
+	UnsignedInt mask = (UnsignedInt)(intptr_t)userData;
 
 	if (INI::scanBool(ini->getNextToken()))
 		*s |= mask;
